backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic[aiohttp]>=0.84.0",
    "chromadb>=1.5.2",
    "fastapi>=0.115.0",
    "google-cloud-firestore>=2.19.0",
    "load-dotenv>=0.1.0",
    "openai>=2.26.0",
    "pydantic-settings>=2.13.1",
    "python-frontmatter>=1.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "snowflake-connector-python[pandas]>=4.3.0",
    "sse-starlette>=2.2.0",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
